Cleaning out a thousand lines of cruft.

+++ MSpectrum.h
@@ -52,7 +52,7 @@ public:
   char**          kojakSparseArray;
   int             kojakBins;
   
-  float lowScore;
+  double lowScore;
 
   int cc;
   int sc;
